A few fixes and SizingManager

+++ src/EnergyPlus/WindTurbine.cc
@@ -796,7 +796,7 @@ namespace WindTurbine {
                 while (statFile.good()) { // end of file
                     auto lineIn = statFile.readLine();
                     // reconcile line with different versions of stat file
-                    int lnPtr = index(lineIn.data, "Wind Speed");
+                    size_t lnPtr = index(lineIn.data, "Wind Speed");
                     if (lnPtr == std::string::npos) continue;
                     // have hit correct section.
                     while (statFile.good()) { // find daily avg line
